Fix state of score on each assessment
For the assessment page. Made an assessment state that contains all answers on a task from the inspera dataset. Default score is `null`, but it should be integrated with the local storage system coming up soon.

As of now there is a **problem** with reapplying the state of score in the pointbox when moving back to previous scored answers.