Enable reAssessments
Now, reAssessments are appended to the end of assessments. You can keep
track of which assessments that are reassessments by either checking for
two assessments with same answers or using the variable
`numberOfAnswers` (which refers to unique answers, not assessments).

I sliced the answerlist so that the user doesn't have to go through too
many answers to se the reassessments.