Create completion page
A simple page is created viewing the grades of each student. The cards are clickable, but the page it leads to is not created yet.
One check to get the tasks is commented out and the tasknumbers to be viewed are set manually so it is easier to test (aka you don't need to assess ALL of the tasks to view completion page with grades).

There is one warning in the console: Warning: Expected server HTML to contain a matching <div> in <div>.
A solutions is probably here: https://localcoder.org/react-16-warning-expected-server-html-to-contain-a-matching-div-in-body, but it took me some time trying to understand so I'll rather prioritize other development.