redesign consistency box
I just moved the collapse further up so that the user don't get that much information at once.
Also I changed the text to contain more "din" so that it might be more understandeble whom's assessements that will be
re-evaluated.
And I moved the toggle botton next to the text rather than underneath.